pyproject.toml
[project]
name = "habit_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "emoji>=2.14.1",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "firebase-admin>=6.8.0",
    "google-api-python-client>=2.167.0",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.54",
    "langchain-google-genai>=2.1.3",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.3.31",
    "mcp>=1.6.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "requests>=2.32.3",
    "ruff>=0.11.6",
    "streamlit>=1.44.1",
    "streamlit-calendar>=1.2.1",
    "uvicorn>=0.34.2",
]

[tool.ruff]
indent-width = 2
line-length = 120
